Script improvements

+++ configSystem.sh
@@ -1,7 +1,12 @@
 #!/usr/bin/env bash
 
-USER=mamutal91
-HOST=odin
+clear
+
+read -p "You user? [ enter = mamutal91 ]: " username
+read -p "You hostname? [ enter = odin ]: " hostname
+[[ -z $username ]] && username=mamutal91 || username=$username
+[[ -z $hostname ]] && hostname=odin || hostname=$hostname
+echo -e "\nUSER: $username\nHOST: $hostname\n"
 
 ln -s /hostlvm /run/lvm
 
@@ -19,7 +24,7 @@ sed -i "s/block/block encrypt lvm2/g" /etc/mkinitcpio.conf
 mkinitcpio -P
 
 echo "Config sudoers"
-sed -i "s/root ALL=(ALL) ALL/root ALL=(ALL) NOPASSWD: ALL\n$USER ALL=(ALL) NOPASSWD:ALL/g" /etc/sudoers
+sed -i "s/root ALL=(ALL) ALL/root ALL=(ALL) NOPASSWD: ALL\n$username ALL=(ALL) NOPASSWD:ALL/g" /etc/sudoers
 
 # systemd
 sed -i "s/#HandleLidSwitch=suspend/HandleLidSwitch=ignore/g" /etc/systemd/logind.conf
@@ -47,11 +52,11 @@ grub-install --target=x86_64-efi --efi-directory=/boot --bootloader-id=grub
 grub-mkconfig -o /boot/grub/grub.cfg
 
 echo "Add my user"
-useradd -m -G wheel -s /bin/bash $USER
-mkdir -p /home/$USER
+useradd -m -G wheel -s /bin/bash $username
+mkdir -p /home/$username
 
-if [[ $USER == "mamutal91" ]]; then
-  git clone https://github.com/mamutal91/dotfiles /home/$USER/.dotfiles
+if [[ $username == "mamutal91" ]]; then
+  git clone https://github.com/mamutal91/dotfiles /home/$username/.dotfiles
 fi
 
 echo "Set locale, zone and keymap console"
@@ -65,22 +70,27 @@ locale-gen
 sudo ln -sf /usr/share/zoneinfo/America/Sao_Paulo /etc/localtime
 hwclock --systohc
 
-echo $HOST > /etc/hostname
+echo $hostname > /etc/hostname
 
 # Mount HDD storage
-echo "" >> /etc/crypttab
-echo "storage UUID=$(blkid /dev/sda1 | awk -F '"' '{print $2}') /root/keyfile luks" >> /etc/crypttab
-echo "" >> /etc/fstab
-echo "/dev/mapper/storage  /media/storage     ext4    defaults        0       2" /etc/fstab
-dd if=/dev/urandom of=/root/keyfile bs=1024 count=4
-chmod 0400 /root/keyfile
-cryptsetup -v luksAddKey /dev/sda1 /root/keyfile
+if [[ $username == mamutal91 ]]; then
+  mkdir -p /mnt/media/storage
+  echo "" >> /etc/crypttab
+  echo "storage UUID=$(blkid /dev/sda1 | awk -F '"' '{print $2}') /root/keyfile luks" >> /etc/crypttab
+  echo "" >> /etc/fstab
+  echo "# Storage" >> /etc/fstab
+  echo "/dev/mapper/storage  /media/storage     ext4    defaults        0       2" >> /etc/fstab
+  dd if=/dev/urandom of=/root/keyfile bs=1024 count=4
+  chmod 0400 /root/keyfile
+  cryptsetup -v luksAddKey /dev/sda1 /root/keyfile
+fi
 
 # Services
 systemctl disable NetworkManager
 systemctl enable dhcpcd
 systemctl enable iwd
 
+clear
 echo "Set passwords"
-passwd $USER
+passwd $username
 passwd root
